Added ability to attach another spine sprite to a slot on a spine sprite.
If you have a sprite X, you can attach sprite Y to one of its slots by calling:

X.attachSpineSprite("slotName", Y);

If something was currently rendering on that slot (i.e. a Mesh or a Region), then it will no longer
be rendered while the Spine sprite is attached. You are swapping it out.

To return back to whatever the slot would normally be rendering (if you hadn't swapped anything),
simply remove the attachment:

X.detachSpineSprite("slotName");

You are responsible for ensuring the attached sprite is not destroyed during the time it is being
used as an attachment.